+++ src/main/java/com/secuconnect/client/model/PaymentContractsDTORequestIdAdditionalDataCreateCard.java
@@ -7,31 +7,14 @@ import java.util.Objects;
  * Create card input
  */
 public class PaymentContractsDTORequestIdAdditionalDataCreateCard {
-  @SerializedName("card_scheme")
-  protected String cardScheme = null;
-
   @SerializedName("app_language")
   protected String appLanguage = null;
 
-  @SerializedName("passport_id")
-  protected String passportId = null;
+  @SerializedName("card_type")
+  protected String cardType = null;
 
-  public PaymentContractsDTORequestIdAdditionalDataCreateCard cardScheme(String cardScheme) {
-    this.cardScheme = cardScheme;
-    return this;
-  }
-
-   /**
-   * Credit card schema
-   * @return cardScheme
-  **/
-  public String getCardScheme() {
-    return cardScheme;
-  }
-
-  public void setCardScheme(String cardScheme) {
-    this.cardScheme = cardScheme;
-  }
+  @SerializedName("identifier")
+  protected String identifier = null;
 
   public PaymentContractsDTORequestIdAdditionalDataCreateCard appLanguage(String appLanguage) {
     this.appLanguage = appLanguage;
@@ -50,21 +33,38 @@ public class PaymentContractsDTORequestIdAdditionalDataCreateCard {
     this.appLanguage = appLanguage;
   }
 
-  public PaymentContractsDTORequestIdAdditionalDataCreateCard passportId(String passportId) {
-    this.passportId = passportId;
+  public PaymentContractsDTORequestIdAdditionalDataCreateCard cardType(String cardType) {
+    this.cardType = cardType;
     return this;
   }
 
    /**
-   * Passport ID
-   * @return passportId
+   * Card Type
+   * @return cardType
   **/
-  public String getPassportId() {
-    return passportId;
+  public String getCardType() {
+    return cardType;
   }
 
-  public void setPassportId(String passportId) {
-    this.passportId = passportId;
+  public void setCardType(String cardType) {
+    this.cardType = cardType;
+  }
+
+  public PaymentContractsDTORequestIdAdditionalDataCreateCard identifier(String identifier) {
+    this.identifier = identifier;
+    return this;
+  }
+
+   /**
+   * Identifier - required if card_type is &#x27;plastic&#x27;
+   * @return identifier
+  **/
+  public String getIdentifier() {
+    return identifier;
+  }
+
+  public void setIdentifier(String identifier) {
+    this.identifier = identifier;
   }
 
   @Override
@@ -76,23 +76,23 @@ public class PaymentContractsDTORequestIdAdditionalDataCreateCard {
       return false;
     }
     PaymentContractsDTORequestIdAdditionalDataCreateCard paymentContractsDTORequestIdAdditionalDataCreateCard = (PaymentContractsDTORequestIdAdditionalDataCreateCard) o;
-    return Objects.equals(this.cardScheme, paymentContractsDTORequestIdAdditionalDataCreateCard.cardScheme) &&
-        Objects.equals(this.appLanguage, paymentContractsDTORequestIdAdditionalDataCreateCard.appLanguage) &&
-        Objects.equals(this.passportId, paymentContractsDTORequestIdAdditionalDataCreateCard.passportId);
+    return Objects.equals(this.appLanguage, paymentContractsDTORequestIdAdditionalDataCreateCard.appLanguage) &&
+        Objects.equals(this.cardType, paymentContractsDTORequestIdAdditionalDataCreateCard.cardType) &&
+        Objects.equals(this.identifier, paymentContractsDTORequestIdAdditionalDataCreateCard.identifier);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(cardScheme, appLanguage, passportId);
+    return Objects.hash(appLanguage, cardType, identifier);
   }
 
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
     sb.append("class PaymentContractsDTORequestIdAdditionalDataCreateCard {\n");
-    sb.append("    cardScheme: ").append(toIndentedString(cardScheme)).append("\n");
     sb.append("    appLanguage: ").append(toIndentedString(appLanguage)).append("\n");
-    sb.append("    passportId: ").append(toIndentedString(passportId)).append("\n");
+    sb.append("    cardType: ").append(toIndentedString(cardType)).append("\n");
+    sb.append("    identifier: ").append(toIndentedString(identifier)).append("\n");
     sb.append("}");
     return sb.toString();
   }